+++ pkg/utils/instance.go
@@ -1,0 +1,7 @@
+package utils
+
+import "fmt"
+
+func GenerateInstanceName(prefix string, nodeClaimName string) string {
+	return fmt.Sprintf("%s-%s", prefix, nodeClaimName)
+}