+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 
 - Add Karpenter deployment manifests
 - provider: drop clusterName (unused attribute)
+- cloudprovider: use clusterID instead of clusterName
+- Add EXOSCALE_COMPUTE_INSTANCE_PREFIX configuration
 
 0.0.1
 ------